+++ RELEASE_NOTES.md
@@ -14,6 +14,7 @@ This serves two purposes:
 
 ### Changed
 - internal: The HydeKernel has been refactored to move related logic to service classes. This does not change the end usage as the Hyde facade still works the same
+- `DiscoveryService::getSourceFileListForModel()` now throws an exception instead of returning false when given an invalid model class
 
 ### Deprecated
 - for soon-to-be removed features.